Export component to different file
